Rename Copy to CopyButton and avoid circular imports
